# openapi.yml
openapi: 3.1.0
info:
  description: The same-origin contract consumed by the BeyondPilot web application.
  title: BeyondPilot API
  version: 1.0.0
servers:
- description: Same origin as the web application
  url: /
...
paths:
...
  /api/organization/mine:
    get:
      operationId: getMyOrganization
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyOrganization"
          description: Where the caller stands.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
      security:
      - session: []
      summary: "The caller's organization, or their ways into one"
      tags:
      - My organization
    put:
      operationId: saveMyOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveOrganization"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
          description: "The organization as saved, with its new version."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A member is not valid.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller belongs to no organization, or is not an owner\
            \ of it."
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The organization changed since it was read.
      security:
      - session: []
      summary: Save the profile of the caller's organization
      tags:
      - My organization
  /api/organization/mine/auto-join:
    put:
      operationId: changeOrganizationAutoJoin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoJoin"
        required: true
      responses:
        "204":
          description: The setting is saved.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The value is missing.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller belongs to no organization, or is not an owner\
            \ of it."
      security:
      - session: []
      summary: "Let addresses on the organization's domain join without asking, or\
        \ stop that"
      tags:
      - My organization
  /api/organization/mine/invitations:
    post:
      operationId: inviteOrganizationMember
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMember"
        required: true
      responses:
        "204":
          description: The invitation is open and the address was told.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A member is not valid.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller belongs to no organization, or is not an owner\
            \ of it."
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The address already belongs to the organization, or already\
            \ holds an open invitation."
      security:
      - session: []
      summary: Ask an address to join the caller's organization
      tags:
      - My organization
  /api/organization/mine/invitations/{id}/revoke:
    post:
      operationId: revokeOrganizationInvitation
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The invitation is closed.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller belongs to no organization, or is not an owner\
            \ of it."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The invitation is not open.
      security:
      - session: []
      summary: Take back an open invitation
      tags:
      - My organization
  /api/organization/mine/job-title:
    put:
      operationId: changeMyJobTitle
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobTitle"
        required: true
      responses:
        "204":
          description: The job title is saved.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The job title is too long.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller belongs to no organization.
      security:
      - session: []
      summary: Set what the caller does in their organization
      tags:
      - My organization
  /api/organization/mine/members:
    get:
      operationId: listMyOrganizationMembers
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMembers"
          description: "The members, and for an owner the open invitations and requests."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller belongs to no organization.
      security:
      - session: []
      summary: Who belongs to the caller's organization
      tags:
      - My organization
  /api/organization/mine/members/{accountId}/remove:
    post:
      operationId: removeOrganizationMember
      parameters:
      - in: path
        name: accountId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The person no longer belongs to the organization.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller belongs to no organization, or removes someone\
            \ else without being an owner."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The person does not belong to the organization.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The removal would leave the organization without an owner.
      security:
      - session: []
      summary: Take a person out of the organization; a member may only take themselves
        out
      tags:
      - My organization
  /api/organization/mine/members/{accountId}/role:
    put:
      operationId: changeOrganizationMemberRole
      parameters:
      - in: path
        name: accountId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeMemberRole"
        required: true
      responses:
        "204":
          description: The person has the role.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The role is not valid.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller belongs to no organization, or is not an owner\
            \ of it."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The person does not belong to the organization.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The change would leave the organization without an owner.
      security:
      - session: []
      summary: "Make a member an owner, or an owner a member"
      tags:
      - My organization
  /api/organization/mine/requests/{id}/approve:
    post:
      operationId: approveJoinRequest
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The person is a member.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller belongs to no organization, or is not an owner\
            \ of it."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The request is not open.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The person joined another organization in the meantime.
      security:
      - session: []
      summary: Let a person who asked join
      tags:
      - My organization
  /api/organization/mine/requests/{id}/decline:
    post:
      operationId: declineJoinRequest
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The request is closed.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller belongs to no organization, or is not an owner\
            \ of it."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The request is not open.
      security:
      - session: []
      summary: Decline a request to join
      tags:
      - My organization
...
components:
  schemas:
...
    AutoJoin:
      type: object
      description: Whether an address on the organization's domain joins without asking.
      properties:
        autoJoin:
          type: boolean
      required:
      - autoJoin
    ChangeMemberRole:
      type: object
      properties:
        role:
          type: string
          enum:
          - owner
          - member
          pattern: owner|member
      required:
      - role
...
    InviteMember:
      type: object
      description: An address asked to join the caller's organization.
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          minLength: 0
        role:
          type: string
          enum:
          - owner
          - member
          pattern: owner|member
      required:
      - email
      - role
    JobTitle:
      type: object
      description: What the caller does in their organization; null or blank removes
        it.
      properties:
        jobTitle:
          type:
          - string
          - "null"
          maxLength: 120
          minLength: 0
...
    MyOrganization:
      type: object
      description: "Where the caller stands: the organization they belong to, or their\
        \ ways in."
      properties:
        invitations:
          type: array
          description: The open invitations to the caller's address.
          items:
            $ref: "#/components/schemas/OrganizationInvitation"
        jobTitle:
          type:
          - string
          - "null"
        organization:
          type:
          - object
          - "null"
          $ref: "#/components/schemas/Organization"
          description: The organization the caller belongs to.
        request:
          type:
          - object
          - "null"
          $ref: "#/components/schemas/OrganizationJoinRequest"
          description: The request the caller waits on.
        role:
          type:
          - string
          - "null"
          description: What the caller is in it.
          enum:
          - owner
          - member
        suggestion:
          type:
          - object
          - "null"
          $ref: "#/components/schemas/OrganizationMatch"
          description: "The organization of the caller's email domain, when they belong\
            \ to none."
      required:
      - invitations
...
    Organization:
      type: object
      description: "An organization as the people who belong to it, and operators,\
        \ see it."
      properties:
        autoJoin:
          type: boolean
          description: Whether an address on the domain joins at once; otherwise it
            asks the owners.
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
        createdAt:
          type: string
          format: date-time
        decisionMessage:
          type:
          - string
          - "null"
          description: What the operator wrote to the owners with the refusal.
        decisionReason:
          type:
          - string
          - "null"
          description: Why it was last refused.
          enum:
          - duplicate
          - not_a_real_organization
          - incomplete
          - out_of_scope
          - other
        description:
          type:
          - string
          - "null"
        emailDomain:
          type:
          - string
          - "null"
          description: The domain whose addresses may join; null when it was made
            from a public mail address.
        id:
          type: string
          format: uuid
        name:
          type: string
        roles:
          type: array
          description: "What it does here: `provider` lists AI solutions, `enterprise`\
            \ posts use cases."
          items:
            type: string
        slug:
          type: string
        status:
          type: string
          description: GenAI Fund's review of the organization.
          enum:
          - pending
          - approved
          - rejected
        teamSize:
          type:
          - string
          - "null"
          description: Null is unknown.
          enum:
          - just_me
          - 2_9
          - 10_49
          - 50_99
          - 100_499
          - 500_999
          - 1000_4999
          - 5000_plus
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
        version:
          type: integer
          format: int64
          description: "Sent back with a save, which is refused when the organization\
            \ changed since."
        website:
          type:
          - string
          - "null"
      required:
      - autoJoin
      - createdAt
      - id
      - name
      - roles
      - slug
      - status
      - type
      - version
...
    OrganizationMembers:
      type: object
      description: "Who belongs to an organization, owners first. Invitations and\
        \ requests are empty for a caller who is not an owner."
      properties:
        invitations:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationInvitation"
        members:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationMember"
        requests:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationJoinRequest"
      required:
      - invitations
      - members
      - requests
...
    Problem:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          description: Stable failure code for module failures and request validation;
            clients branch on it.
          maxLength: 63
          pattern: "^[A-Z][A-Z0-9_]+[A-Z0-9]$"
        detail:
          type: string
          description: Safe human-readable explanation; never branch on it.
        errors:
          type: array
          description: "Request validation only: one entry per violation. A rejected\
            \ value is never echoed."
          items:
            type: object
            additionalProperties: false
            properties:
              code:
                type: string
                description: "Constraint code, for example `NOT_BLANK` or `SIZE`."
                pattern: "^[A-Z][A-Z0-9_]+[A-Z0-9]$"
              detail:
                type: string
                description: Safe fallback text; clients translate by code.
              params:
                type: object
                additionalProperties: false
                description: Numeric bounds of the constraint; omitted when it has
                  none.
                properties:
                  max:
                    type: number
                  min:
                    type: number
              pointer:
                type: string
                description: "JSON Pointer to the offending member, for example `#/title`."
            required:
            - code
            - detail
            - pointer
        instance:
          type: string
          format: uri-reference
          description: Request path that produced the problem.
        requestId:
          type: string
          format: uuid
          description: Identifier of the request in the server logs and the `X-Request-Id`
            header.
        status:
          type: integer
          format: int32
          description: HTTP status code.
        title:
          type: string
          description: Short summary of the problem kind; never branch on it.
        type:
          type: string
          format: uri
          description: '`urn:beyondpilot:failure:<code in kebab case>` when `code`
            is present; omitted means `about:blank`.'
      required:
      - requestId
      - status
      - title
...
    SaveOrganization:
      type: object
      description: The profile of an organization as its Profile screen holds it.
      properties:
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
          pattern: "[A-Z]{2}"
        description:
          type:
          - string
          - "null"
          maxLength: 2000
          minLength: 0
        name:
          type: string
          maxLength: 120
          minLength: 0
        roles:
          type: array
          description: One or both of `provider` and `enterprise`.
          items:
            type: string
            pattern: provider|enterprise
          maxItems: 2
          minItems: 1
        teamSize:
          type:
          - string
          - "null"
          pattern: just_me|2_9|10_49|50_99|100_499|500_999|1000_4999|5000_plus
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
          pattern: company|builder_team|independent_builder|other
        version:
          type: integer
          format: int64
          description: The version the screen read.
        website:
          type:
          - string
          - "null"
          maxLength: 300
          minLength: 0
          pattern: "https?://[^\\s]+"
      required:
      - name
      - roles
      - type
      - version
...
  securitySchemes:
    session:
      description: "The session cookie set by signing in. Requests that change state\
        \ also send `X-BeyondPilot-CSRF: 1`."
      in: cookie
      name: BEYONDPILOT_SESSION
      type: apiKey